se hace el cambio en plazo de float a int

+++ src/view/console/consola_credito.py
@@ -8,7 +8,7 @@ try:
     print("Este programa le permite calcular la cuota a pagar por un credito educativo")
     monto_credito = float(input("Monto del credito: "))
     tasa = float(input("Tasa de interes mensual del credito: ")) / 100
-    plazo = float(input("Numero de cuotas en que va a pagar el credito: "))
+    plazo = int(input("Numero de cuotas en que va a pagar el credito: "))
 
     cuota = round(logica_credito.calcular_cuota(monto_credito, tasa, plazo), 2)
     total_abonos = round(logica_credito.calcular_total_abonos(monto_credito, tasa, plazo), 2)
